+++ tests/test_ribocount.py
@@ -1,4 +1,3 @@
-import os
 import logging
 import unittest
 
@@ -19,8 +18,13 @@ class RiboCountTestCase(unittest.TestCase):
 
     def test_get_ribo_counts(self):
         """Get read counts upstream of the longest ORF"""
-        with ribocore.open_pysam_file(CFG.RIBO_FILE, ftype='bam') as f:
-            counts, reads = ribocore.get_ribo_counts(ribo_fileobj=f, transcript_name=CFG.TRANSCRIPT_NAME)
+        parser = ribocount.create_parser()
+        args = parser.parse_args(['-b', CFG.RIBO_FILE, '-f', CFG.UNRELATED_FASTA, '-l', '0', '-s', '0'])
+        with ribocore.open_pysam_file(args.ribo_file, ftype='bam') as f:
+            # pass transcript name explicitly as it is not one of the command
+            # line options
+            counts, reads = ribocore.get_ribo_counts(ribo_fileobj=f, transcript_name=CFG.TRANSCRIPT_NAME,
+                                                     read_lengths=args.read_lengths, read_offsets=args.read_offsets)
 
             # 529 is the start position of the longest orf for the test transcript
             counts_upstream, reads_upstream = ribocore.filter_ribo_counts(counts=counts, orf_start=529)